Improve the App configuration files order and rename app/Config/ClassAliases.php as App.php
